% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.r
\name{boxcox}
\alias{boxcox}
\title{Box-Cox Transform}
\usage{
boxcox(x, lambda, plot = FALSE)
}
\arguments{
\item{x}{The vector to be transformed. Note that x must be a positive vector.}

\item{lambda}{The power used in the transformation}

\item{plot}{Display a plot of \code{x} vs the output? Use logical.
\code{FALSE} by default.}
}
\value{
A vector that is the transformed results of \code{x}. Plots the data if \code{plot} is \code{TRUE}.
}
\description{
This function applies the Box-Cox transformation on a vector.
}
